requirements.txt
absl-py==2.1.0
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
chronology==1.1
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
contourpy==1.2.0
coverage==7.5.0
cycler==0.12.1
dame-flame==0.61
dash==2.16.0
dash-core-components==2.0.0
dash-cytoscape==1.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
et-xmlfile==1.1.0
executing==2.0.1
fake-useragent==1.4.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.47.2
gast==0.5.4
gevent==24.2.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
interpret==0.5.1
interpret-core==0.5.1
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
ml-dtypes==0.3.2
multiprocess==0.70.16
namex==0.0.7
navigation==0.1
nest-asyncio==1.5.9
numba==0.59.0
numpy==1.26.3
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.19.0
pluggy==1.4.0
progressbar2==4.3.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pygam==0.9.1
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.0.0
python-dateutil==2.8.2
python-utils==3.8.2
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
retrying==1.3.4
rich==13.7.1
ruff==0.3.7
SALib==1.4.8
scikit-learn==1.4.1.post1
scipy==1.11.4
seaborn==0.13.1
selenium==4.17.2
shap==0.44.1
six==1.16.0
slicer==0.0.7
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.0.3
xlrd==2.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
Pillow==10.2.0